Remove unused sys.utils
