--- src/sys/utils/utils.go
@@ -1,35 +1,0 @@
-package utils
-
-import (
-	"github.com/pkg/errors"
-	"github.com/spf13/viper"
-)
-
-func ValueOrDefault[T any](input *T, defaultValue T) T {
-	if input == nil {
-		return defaultValue
-	}
-	return *input
-}
-
-func EnvOrDefault[T any](envKey string, defaultValue T) T {
-	if viper.IsSet(envKey) {
-		result, ok := viper.Get(envKey).(T)
-		if ok {
-			return result
-		}
-	}
-	return defaultValue
-}
-
-func ArgStringD(arg, def string) string {
-	v := arg
-	if arg == "" {
-		v = def
-	}
-	return v
-}
-
-func MakeValidationError(field, rule string) error {
-	return errors.Errorf("%s failed on validation '%s'", field, rule)
-}
