--- src/sys/utils/utils_test.go
@@ -1,32 +1,0 @@
-package utils
-
-import (
-	"testing"
-
-	"github.com/icrowley/fake"
-	"github.com/stretchr/testify/assert"
-)
-
-func Test_ValueOrDefault_Input(t *testing.T) {
-	input := "foo"
-	result := ValueOrDefault[string](&input, "bar")
-	assert.Equal(t, input, result)
-}
-
-func Test_ValueOrDefault_Default(t *testing.T) {
-	result := ValueOrDefault[string](nil, "bar")
-	assert.Equal(t, "bar", result)
-}
-
-func TestArgStringD_Provided(t *testing.T) {
-	a := fake.Word()
-	d := fake.Word()
-	v := ArgStringD(a, d)
-	assert.Equal(t, a, v, "failed to return provided arg")
-}
-
-func TestArgStringD_Default(t *testing.T) {
-	d := fake.Word()
-	v := ArgStringD("", d)
-	assert.Equal(t, d, v, "failed to return default arg")
-}
