+++ src/utils.go
@@ -1,11 +1,21 @@
 package goat
 
-import "github.com/68696c6c/goat/sys/utils"
+import "github.com/spf13/viper"
 
+// ValueOrDefault returns the value of input or defaultValue if input is nil.
 func ValueOrDefault[T any](input *T, defaultValue T) T {
-	return utils.ValueOrDefault[T](input, defaultValue)
+	if input == nil {
+		return defaultValue
+	}
+	return *input
 }
 
 func EnvOrDefault[T any](envKey string, defaultValue T) T {
-	return utils.EnvOrDefault[T](envKey, defaultValue)
+	if viper.IsSet(envKey) {
+		result, ok := viper.Get(envKey).(T)
+		if ok {
+			return result
+		}
+	}
+	return defaultValue
 }
